+++ ErsatzTV.FFmpeg/Capabilities/Vaapi/VaapiProfile.cs
@@ -1,0 +1,17 @@
+namespace ErsatzTV.FFmpeg.Capabilities.Vaapi;
+
+public class VaapiProfile
+{
+    public const string Mpeg2Simple = "VAProfileMPEG2Simple";
+    public const string Mpeg2Main = "VAProfileMPEG2Main";
+    public const string H264ConstrainedBaseline = "VAProfileH264ConstrainedBaseline";
+    public const string H264Main = "VAProfileH264Main";
+    public const string H264High = "VAProfileH264High";
+    public const string H264MultiviewHigh = "VAProfileH264MultiviewHigh";
+    public const string H264StereoHigh = "VAProfileH264StereoHigh";
+    public const string Vc1Simple = "VAProfileVC1Simple";
+    public const string Vc1Main = "VAProfileVC1Main";
+    public const string Vc1Advanced = "VAProfileVC1Advanced";
+    public const string HevcMain = "VAProfileHEVCMain";
+    public const string HevcMain10 = "VAProfileHEVCMain10";
+}