+++ ErsatzTV.FFmpeg/Capabilities/Vaapi/VaapiProfileEntrypoint.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.FFmpeg.Capabilities.Vaapi;
+
+public record VaapiProfileEntrypoint(string VaapiProfile, string VaapiEntrypoint);