+++ ErsatzTV.FFmpeg/Pipeline/IPipelineBuilderFactory.cs
@@ -8,6 +8,8 @@ public interface IPipelineBuilderFactory
         Option<AudioInputFile> audioInputFile,
         Option<WatermarkInputFile> watermarkInputFile,
         Option<SubtitleInputFile> subtitleInputFile,
+        Option<string> vaapiDriver,
+        Option<string> vaapiDevice,
         string reportsFolder,
         string fontsFolder,
         string ffmpegPath);