+++ ErsatzTV.FFmpeg/Pipeline/QsvPipelineBuilder.cs
@@ -48,8 +48,14 @@ public class QsvPipelineBuilder : SoftwarePipelineBuilder
         PipelineContext context,
         ICollection<IPipelineStep> pipelineSteps)
     {
-        bool canDecode = _hardwareCapabilities.CanDecode(videoStream.Codec, videoStream.PixelFormat);
-        bool canEncode = _hardwareCapabilities.CanEncode(desiredState.VideoFormat, desiredState.PixelFormat);
+        bool canDecode = _hardwareCapabilities.CanDecode(
+            videoStream.Codec,
+            desiredState.VideoProfile,
+            videoStream.PixelFormat);
+        bool canEncode = _hardwareCapabilities.CanEncode(
+            desiredState.VideoFormat,
+            desiredState.VideoProfile,
+            desiredState.PixelFormat);
 
         pipelineSteps.Add(new QsvHardwareAccelerationOption(ffmpegState.VaapiDevice));